refactor: move tar code from git_shell and file_utils to tar_utils
