+++ ggshield/cmd/iac/scan/iac_scan_utils.py
@@ -11,12 +11,8 @@ from ggshield.cmd.utils.common_options import use_json
 from ggshield.core.config import Config
 from ggshield.core.errors import APIKeyCheckError
 from ggshield.core.file_utils import is_filepath_excluded
-from ggshield.core.git_shell import (
-    INDEX_REF,
-    get_filepaths_from_ref,
-    get_staged_filepaths,
-    tar_from_ref_and_filepaths,
-)
+from ggshield.core.git_shell import get_filepaths_from_ref, get_staged_filepaths
+from ggshield.core.tar_utils import INDEX_REF, tar_from_ref_and_filepaths
 from ggshield.core.text_utils import display_error
 from ggshield.verticals.iac.filter import is_iac_file_path
 from ggshield.verticals.iac.output import (
