+++ ggshield/core/file_utils.py
@@ -1,7 +1,5 @@
 import os
 import re
-import tarfile
-from io import BytesIO
 from pathlib import Path, PurePosixPath
 from typing import List, Set, Union
 
@@ -55,10 +53,3 @@ def is_path_binary(path: str) -> bool:
     _, ext = os.path.splitext(path)
     # `[1:]` because `ext` starts with a "." but extensions in `BINARY_EXTENSIONS` do not
     return ext[1:] in BINARY_EXTENSIONS
-
-
-def get_empty_tar() -> bytes:
-    bytes = BytesIO()
-    file = tarfile.open(fileobj=bytes, mode="w:gz")
-    file.close()
-    return bytes.getvalue()
