+++ tests/unit/core/test_file_utils.py
@@ -7,7 +7,8 @@ from typing import Set, Union
 
 import pytest
 
-from ggshield.core.file_utils import get_empty_tar, is_filepath_excluded
+from ggshield.core.file_utils import is_filepath_excluded
+from ggshield.core.tar_utils import get_empty_tar
 
 
 def test_get_empty_tar():
